requirements.txt
certifi==2022.12.7
charset-normalizer==2.1.1
idna==3.4
oauthlib==3.2.2
py4j==0.10.9.5
pyspark==3.3.1
requests==2.28.1
requests-oauthlib==1.3.1
tweepy==4.12.1
urllib3==1.26.13
websocket-client==1.4.2
